Simplify DFM usage. Add profile modules.
This version is backwards incompatible. We're removing the global config file,
this file made DFM awkward to use. You couldn't store the global config in your
dotfiles and so needed to do some manual copying to use dfm correctly which
defeated the purpose. We're instead moving all configuration into the dotfiles
repo itself. This means that to start a new machine all you need to do is
install dfm and run dfm clone then all of your "sub-profiles" or modules will be
handled for you which was the original goal of the multiple locations in the old
global config file.

Additionally, the concept of backends has been removed. This allows me to focus
on using git in useful ways and it is probably the best way to manage dotfiles
anyway. This is an opinionated change but one that cleans up the code and
simplifies it's maintenance and usage.