Add a --name flag to Sync
Now that Modules and Profiles are the same thing we should let you
selectively sync them. If the --name flag is given to sync it will
only sync the profile with the name, if no profile with that name is
found then it will recursively search modules from the current profile
until it finds a module that matches the --name value.