Fix a bug where load_profile called on a new installation failed
current_profile calls sys.exit(1) if there is no current_profile but
because of how Python handles default arguments current_profile was
always being called even if load_profile was called with a name.

This meant that when running dfm clone on a new installation would
fail stating "no current profile found" even though no profiles were
available on the system. Now we default to None and only if we were
explicitly not passed a name do we try to load the current profile
location.