fix: Refactor mapping logic and fix many bugs
Before mapping appeared to work as a result of pure luck and it was not
clear how it behaved in various target_os scenarios.

This refactor separates the logic of mapping from the Profile object and
pushes it down into the mapping object. This includes things like OS
detection for mappings.

As part of this refactor work I've extending the test_linking test cases
to cover a variety of new scenarios. Linking is now simpler and easier
to reason about.