feature: Use git ls-files to discover profile files
This has two major benefits: 1 it is faster, 2 .gitignore can now be
used to ignore files from dfm.

Before we would try to link things that git would be ignoring, this
could lead to all kinds of weird and whacky situations. Now we respect
the .gitignore file but just using git's built-in file listing command.